refactor: extract reviewer purpose info sections into shared components
