+++ src/pages/RiskAnalysisInfoCompilePage/RiskAnalysisInfoCompile.page.tsx
@@ -1,11 +1,16 @@
 import { PurposeQueries } from '@/api/purpose'
-import { PageContainer, SectionContainer } from '@/components/layout/containers'
-import { Link, useNavigate, useParams } from '@/router'
+import { PageContainer } from '@/components/layout/containers'
+import { useNavigate, useParams } from '@/router'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
-import { Button, Grid, Skeleton, Stack } from '@mui/material'
+import { Button, Grid, Stack } from '@mui/material'
 import { useQuery } from '@tanstack/react-query'
-import { InformationContainer } from '@pagopa/interop-fe-commons'
+import {
+  RiskAnalysisPurposeGeneralInfoSection,
+  RiskAnalysisPurposeGeneralInfoSectionSkeleton,
+  RiskAnalysisPurposeLoadEstimateSection,
+  RiskAnalysisPurposeLoadEstimateSectionSkeleton,
+} from '@/components/shared/RiskAnalysisPurposeInfoSections'
 
 const RiskAnalysisInfoCompilePage: React.FC = () => {
   const { t } = useTranslation('purpose', { keyPrefix: 'riskAnalysisInfoCompile' })
@@ -40,59 +45,8 @@ const RiskAnalysisInfoCompilePage: React.FC = () => {
             <RiskAnalysisInfoCompilePageSkeleton />
           ) : (
             <Stack spacing={3}>
-              <SectionContainer title={t('generalInfoSection.label')}>
-                <Stack spacing={3}>
-                  <InformationContainer
-                    label={t('generalInfoSection.eService.label')}
-                    content={
-                      <Link
-                        to="SUBSCRIBE_CATALOG_VIEW"
-                        params={{
-                          eserviceId: purpose.eservice.id,
-                          descriptorId: purpose.eservice.descriptor.id,
-                        }}
-                        target="_blank"
-                      >
-                        {purpose.eservice.name}
-                      </Link>
-                    }
-                  />
-                  <InformationContainer
-                    label={t('generalInfoSection.producer.label')}
-                    content={purpose.eservice.producer.name}
-                  />
-                  <InformationContainer
-                    label={t('generalInfoSection.purposeName.label')}
-                    content={purpose.title}
-                  />
-                  <InformationContainer
-                    label={t('generalInfoSection.purposeDescription.label')}
-                    content={purpose.description}
-                  />
-                  <InformationContainer
-                    label={t('generalInfoSection.isFreeOfCharge.label')}
-                    content={
-                      purpose.isFreeOfCharge
-                        ? t('generalInfoSection.isFreeOfCharge.options.YES')
-                        : t('generalInfoSection.isFreeOfCharge.options.NO')
-                    }
-                  />
-                  {purpose.isFreeOfCharge && (
-                    <InformationContainer
-                      label={t('generalInfoSection.freeOfChargeReason.label')}
-                      content={purpose.freeOfChargeReason || ''}
-                    />
-                  )}
-                </Stack>
-              </SectionContainer>
-              <SectionContainer title={t('loadEstimationSection.label')}>
-                <Stack spacing={3}>
-                  <InformationContainer
-                    label={t('loadEstimationSection.dailyCalls.label')}
-                    content={`${purpose.currentVersion?.dailyCalls ?? purpose.waitingForApprovalVersion?.dailyCalls ?? 1}`}
-                  />
-                </Stack>
-              </SectionContainer>
+              <RiskAnalysisPurposeGeneralInfoSection purpose={purpose} />
+              <RiskAnalysisPurposeLoadEstimateSection purpose={purpose} />
             </Stack>
           )}
         </Grid>
@@ -106,32 +60,11 @@ const RiskAnalysisInfoCompilePage: React.FC = () => {
   )
 }
 
-const RiskAnalysisGeneralInfoSectionSkeleton: React.FC = () => (
-  <SectionContainer title="">
-    <Stack spacing={3}>
-      <Skeleton variant="text" width="40%" height={32} />
-      <Skeleton variant="rectangular" height={56} />
-      <Skeleton variant="rectangular" height={56} />
-      <Skeleton variant="rectangular" height={56} />
-      <Skeleton variant="rectangular" height={56} />
-    </Stack>
-  </SectionContainer>
-)
-
-const RiskAnalysisLoadEstimateSectionSkeleton: React.FC = () => (
-  <SectionContainer title="">
-    <Stack spacing={3}>
-      <Skeleton variant="text" width="30%" height={32} />
-      <Skeleton variant="rectangular" height={56} />
-    </Stack>
-  </SectionContainer>
-)
-
 const RiskAnalysisInfoCompilePageSkeleton: React.FC = () => {
   return (
     <Stack spacing={3}>
-      <RiskAnalysisGeneralInfoSectionSkeleton />
-      <RiskAnalysisLoadEstimateSectionSkeleton />
+      <RiskAnalysisPurposeGeneralInfoSectionSkeleton />
+      <RiskAnalysisPurposeLoadEstimateSectionSkeleton />
     </Stack>
   )
 }
